requirements.txt
Flask~=2.2.2
Werkzeug~=2.2.2
numpy~=1.24.0
Pillow~=9.3.0